adk-agents/pyproject.toml
[project]
name = "image-scoring"
version = "0.1.0"
description = "The image scoring generates an image for a news article, evaluates the image agaionst a set of rules and regenerates the image if necessary ."
authors = [
    {name = "Kishore Jagannath", email = "kishorerj@google.com"}
]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "image_scoring"}, {include = "hyper_local_content"}, {include = "differentiated_materials"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.1"
google-adk = { version = ">=0.5.0", extras = ["eval"] }
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
google-cloud-vision = "^3.4.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { version = ">=1.3.0", extras = ["eval"] }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
pandas = "^2.1.1"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"